requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.13
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-phonenumber-field==8.0.0
django-timezone-field==7.0
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.8
kombu==5.4.0
multidict==6.0.5
phonenumberslite==8.13.44
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
twilio==9.2.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.9.4
